Election score to double, weight factors to NVS-backed PropertyValue<float>, FSD reverse-update.
- Election score changed from uint32_t to double (overflow-safe)
- ELECT_W_BATTERY/ADJACENCY/TENURE/LOWBAT_PENALTY moved from bsp.hpp defines to NvsConfigManager PropertyValue<float> members (NVS-tunable)
- Below-battery-floor is now a score penalty (not disqualifying), pickWinner() simplified
- PropertyValue gains float NVS support (bit-cast to uint32_t), fnvFloat hash helper
- SETTINGS_HASH refactored to readable constexpr computeSettingsHash() function
- FSD updated: Section 3 firmware layers, Section 5 Phase 1 marked complete, Section 6 full file inventory, new Section 7 NvsConfigManager/LedDriver architecture docs

+++ include/mesh_conductor.h
@@ -18,7 +18,7 @@ struct __attribute__((packed)) ElectionScore {
     uint16_t battery_mv;
     uint8_t  peer_count;        // number of peers this node can see
     uint16_t gateway_tenure;    // times this node has been gateway (from NVS)
-    uint32_t score;             // pre-computed score
+    double   score;             // pre-computed score (double for overflow safety)
 };
 
 // --- IMeshRole abstract interface ---
@@ -76,7 +76,7 @@ public:
     static void printStatus();
 
     // Election
-    static uint32_t computeScore();
+    static double computeScore();
     static void runElection();
 
     // Debug
